Fixes #3604 - Merge core/tests into core/impl
