+++ core/impl/src/test/java/cloud/piranha/core/impl/ServletInputStreamTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2002-2023 Manorrock.com. All Rights Reserved.
+ * Copyright (c) 2002-2024 Manorrock.com. All Rights Reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -30,26 +30,103 @@ package cloud.piranha.core.impl;
 import cloud.piranha.core.api.WebApplication;
 import cloud.piranha.core.api.WebApplicationRequest;
 import cloud.piranha.core.api.WebApplicationResponse;
+import jakarta.servlet.ReadListener;
+import jakarta.servlet.ServletInputStream;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import org.junit.jupiter.api.Test;
 
 /**
  * The JUnit tests for the ServletInputStream API.
  * 
  * @author Manfred Riem (mriem@manorrock.com)
  */
-public class ServletInputStreamTest extends cloud.piranha.core.tests.ServletInputStreamTest {
+public class ServletInputStreamTest {
 
-    @Override
-    public WebApplication createWebApplication() {
+    private WebApplication createWebApplication() {
         return new DefaultWebApplication();
     }
 
-    @Override
-    public WebApplicationRequest createWebApplicationRequest() {
+    private WebApplicationRequest createWebApplicationRequest() {
         return new DefaultWebApplicationRequest();
     }
 
-    @Override
-    public WebApplicationResponse createWebApplicationResponse() {
+    private WebApplicationResponse createWebApplicationResponse() {
         return new DefaultWebApplicationResponse();
+    }
+
+    /**
+     * Test isFinished method.
+     * 
+     * @throws Exception when a serious error occurs.
+     */
+    void testIsFinished() throws Exception {
+        WebApplicationRequest request = createWebApplicationRequest();
+        ServletInputStream inputStream = request.getInputStream();
+        assertFalse(inputStream.isFinished());
+    }
+    
+    /**
+     * Test isReady method.
+     * 
+     * @throws Exception when a serious error occurs.
+     */
+    void testIsReady() throws Exception {
+        WebApplicationRequest request = createWebApplicationRequest();
+        ServletInputStream inputStream = request.getInputStream();
+        assertFalse(inputStream.isReady());
+    }
+    
+    /**
+     * Test read method.
+     * 
+     * @throws Exception when a serious error occurs.
+     */
+    @Test
+    void testRead() throws Exception {
+        WebApplicationRequest request = createWebApplicationRequest();
+        ServletInputStream inputStream = request.getInputStream();
+        request.getWebApplicationInputStream()
+                .setInputStream(new ByteArrayInputStream("a".getBytes()));
+        int character = inputStream.read();
+        assertEquals('a', character);
+    }
+
+    /**
+     * Test setReadListener method.
+     * 
+     * @throws Exception when a serious error occurs.
+     */
+    @Test
+    void testSetReadListener() throws Exception {
+        WebApplication webApplication = createWebApplication();
+        WebApplicationRequest request = createWebApplicationRequest();
+        request.setAsyncSupported(true);
+        request.setWebApplication(webApplication);
+        WebApplicationResponse response = createWebApplicationResponse();
+        response.setWebApplication(webApplication);
+        webApplication.linkRequestAndResponse(request, response);
+        request.startAsync();
+        ServletInputStream inputStream = request.getInputStream();
+        inputStream.setReadListener(new ReadListener() {
+            @Override
+            public void onDataAvailable() throws IOException {
+                throw new UnsupportedOperationException("Not supported yet.");
+            }
+
+            @Override
+            public void onAllDataRead() throws IOException {
+                throw new UnsupportedOperationException("Not supported yet.");
+            }
+
+            @Override
+            public void onError(Throwable t) {
+                throw new UnsupportedOperationException("Not supported yet.");
+            }
+        });
+        assertNotNull(request.getWebApplicationInputStream().getReadListener());
     }
 }
