+++ core/pom.xml
@@ -18,7 +18,6 @@
     <modules>
         <module>api</module>
         <module>impl</module>
-        <module>tests</module>
     </modules>
 
     <dependencyManagement>
